requirements.txt
Flask
pandas
python-dotenv
firebase-admin
google-generativeai
openpyxl
gunicorn